+++ fbw-a32nx/src/systems/fmgc/src/flightplanning/A320AircraftConfig.ts
@@ -142,8 +142,9 @@ const engineModelParams: EngineModelParameters = {
 };
 
 const fmsSymbolConfig: FMSymbolsConfig = {
-  publishDepartureIdent: false,
-  showRnpArLabel: false,
+  showSidName: false,
+  rnpArNaming: false,
+  approachIdentSize: 9,
 };
 
 export const A320AircraftConfig: AircraftConfig = {